♻️ Bound fuzz inputs instead of rejecting them

+++ test/fuzz/cl-managers/compounders/CompounderUniswapV4/ExecuteAction.fuzz.t.sol
@@ -69,7 +69,7 @@ contract ExecuteAction_CompounderUniswapV4_Fuzz_Test is CompounderUniswapV4_Fuzz
         // And: maxSwapFee is smaller or equal to 1e18.
         maxSwapFee = uint64(bound(maxSwapFee, 0, 1e18));
 
-        // And info is set.
+        // And: info is set.
         vm.prank(account.owner());
         compounder.setAccountInfo(
             address(account), initiator, maxClaimFee, maxSwapFee, MAX_TOLERANCE, MIN_LIQUIDITY_RATIO, ""
@@ -101,7 +101,7 @@ contract ExecuteAction_CompounderUniswapV4_Fuzz_Test is CompounderUniswapV4_Fuzz
         // And: maxSwapFee is smaller or equal to 1e18.
         maxSwapFee = uint64(bound(maxSwapFee, 0, 1e18));
 
-        // And info is set.
+        // And: info is set.
         vm.prank(account.owner());
         compounder.setAccountInfo(
             address(account), initiator, maxClaimFee, maxSwapFee, MAX_TOLERANCE, MIN_LIQUIDITY_RATIO, ""
@@ -217,7 +217,7 @@ contract ExecuteAction_CompounderUniswapV4_Fuzz_Test is CompounderUniswapV4_Fuzz
         deal(address(token1), address(compounder), initiatorParams.amount1, true);
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, type(uint8).max, type(uint48).max));
+        feeSeed = bound(feeSeed, type(uint8).max, type(uint48).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
@@ -431,7 +431,7 @@ contract ExecuteAction_CompounderUniswapV4_Fuzz_Test is CompounderUniswapV4_Fuzz
         deal(address(token1), address(compounder), initiatorParams.amount1, true);
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, type(uint8).max, type(uint48).max));
+        feeSeed = bound(feeSeed, type(uint8).max, type(uint48).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
@@ -553,7 +553,7 @@ contract ExecuteAction_CompounderUniswapV4_Fuzz_Test is CompounderUniswapV4_Fuzz
         deal(address(token1), address(compounder), initiatorParams.amount1, true);
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, type(uint8).max, type(uint48).max));
+        feeSeed = bound(feeSeed, type(uint8).max, type(uint48).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
