+++ test/fuzz/cl-managers/rebalancers/RebalancerUniswapV3/ExecuteAction.fuzz.t.sol
@@ -68,7 +68,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
         // And: maxSwapFee is smaller or equal to 1e18.
         maxSwapFee = uint64(bound(maxSwapFee, 0, 1e18));
 
-        // And info is set.
+        // And: info is set.
         vm.prank(account.owner());
         rebalancer.setAccountInfo(
             address(account),
@@ -108,7 +108,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
         // And: maxSwapFee is smaller or equal to 1e18.
         maxSwapFee = uint64(bound(maxSwapFee, 0, 1e18));
 
-        // And info is set.
+        // And: info is set.
         vm.prank(account.owner());
         rebalancer.setAccountInfo(
             address(account),
@@ -636,7 +636,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
             uint128(bound(initiatorParams.amountOut1, 0, (amount1 + initiatorParams.amountIn1) / 2));
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint56).max));
+        feeSeed = bound(feeSeed, 0, type(uint56).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
@@ -762,7 +762,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
             uint128(bound(initiatorParams.amountOut1, 0, (amount1 + initiatorParams.amountIn1) / 2));
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint56).max));
+        feeSeed = bound(feeSeed, 0, type(uint56).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
@@ -889,7 +889,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
         deal(address(token1), address(rebalancer), initiatorParams.amountIn1, true);
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint56).max));
+        feeSeed = bound(feeSeed, 0, type(uint56).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
@@ -1017,7 +1017,7 @@ contract ExecuteAction_RebalancerUniswapV3_Fuzz_Test is RebalancerUniswapV3_Fuzz
         deal(address(token1), address(rebalancer), initiatorParams.amountIn1, true);
 
         // And: Position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint56).max));
+        feeSeed = bound(feeSeed, 0, type(uint56).max);
         generateFees(feeSeed, feeSeed);
 
         // And: account is set.
