+++ test/fuzz/cl-managers/yield-claimers/YieldClaimerUniswapV3/ExecuteAction.fuzz.t.sol
@@ -83,7 +83,7 @@ contract ExecuteAction_YieldClaimerUniswapV3_Fuzz_Test is YieldClaimerUniswapV3_
         // forge-lint: disable-end(erc20-unchecked-transfer)
 
         // And: position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint48).max));
+        feeSeed = bound(feeSeed, 0, type(uint48).max);
         generateFees(feeSeed, feeSeed);
         (uint256 fee0, uint256 fee1) = getFeeAmountsV3(position.id);
 
@@ -181,7 +181,7 @@ contract ExecuteAction_YieldClaimerUniswapV3_Fuzz_Test is YieldClaimerUniswapV3_
         // forge-lint: disable-end(erc20-unchecked-transfer)
 
         // And: position has fees.
-        feeSeed = uint256(bound(feeSeed, 0, type(uint48).max));
+        feeSeed = bound(feeSeed, 0, type(uint48).max);
         generateFees(feeSeed, feeSeed);
         (uint256 fee0, uint256 fee1) = getFeeAmountsV3(position.id);
